Booking About Contacts

+++ main.py
@@ -1,10 +1,11 @@
-from flask import Flask, render_template, url_for, redirect, request
+from flask import Flask, render_template, url_for, redirect, request, flash
 from flask_restful import abort, Api, Resource, reqparse
 from data import db_session
 from flask_login import LoginManager, login_user, logout_user, login_required, current_user
 from data.users import User
 from data.reviews import Review
 from data.albums import Album
+from data.booking import Booking
 from data.forms import *
 import datetime
 from reviews_resourses import ReviewsResource, ReviewsListResource
@@ -57,11 +58,11 @@ def register():
     if form.validate_on_submit():
         if form.password.data != form.password_again.data:
             return render_template('register.html', title='Регистрация', form=form,
-                                   message='Пароли не совпадают')
+                                   message='Пароли не совпадают', albums=get_albums())
         session = db_session.create_session()
         if session.query(User).filter(User.email == form.email.data).first():
             return render_template('register.html', title='Регистрация', form=form,
-                                   message='Пользователь с таким email уже существует')
+                                   message='Пользователь с таким email уже существует', albums=get_albums())
         user = User()
         user.email = form.email.data
         user.set_password(form.password.data)
@@ -108,6 +109,7 @@ def review_form():
         review.date = datetime.date.today()
         session.add(review)
         session.commit()
+        flash('Успех')
         return redirect('/feedback')
     else:
         return render_template('review_form.html', form=form, albums=get_albums())
@@ -119,11 +121,12 @@ def edit_review(id):
     session = db_session.create_session()
     review = session.query(Review).filter(Review.id == id, Review.user_id == current_user.id).first()
     if not review:
-        return render_template('404.html', title='Ошибка 404')
+        abort(404)
     if form.validate_on_submit():
         review.rating = form.rating.data
         review.content = form.content.data
         session.commit()
+        flash('Успех')
         return redirect('/feedback')
     else:
         form.rating.data = review.rating
@@ -139,6 +142,7 @@ def delete_review(id):
     if review:
         session.delete(review)
         session.commit()
+        flash('Успех')
         return redirect('/feedback')
     else:
         return render_template('404.html', title='Ошибка 404', albums=get_albums())
@@ -156,23 +160,28 @@ def albums(album_name):
 
 @app.route('/albums/upload', methods=['GET', 'POST'])
 def upload():
-    if current_user.id == 1:
+    if current_user.is_authenticated and  current_user.id == 1:
         form = AlbumForm()
         session = db_session.create_session()
         if form.validate_on_submit():
             album = session.query(Album).filter(Album.name == form.name.data).first()
             if album:
                 return render_template('album_form.html', title='Загрузить альбом',
-                                       form=form, message='Такой альбом уже существует')
+                                       form=form, message='Такой альбом уже существует', albums=get_albums())
             album = Album()
             album.name = form.name.data
             album.translit_name = ru_to_en(album.name)
             album.lenght = len(form.photos.data)
-            os.mkdir(f'static/albums/{album.translit_name}')
+            try:
+                os.mkdir(f'static/albums/{album.translit_name}')
+            except FileExistsError:
+                return render_template('album_form.html', title='Загрузить альбом',
+                                       form=form, message='Альбом с таким названием уже существует', albums=get_albums())
             for i, file in enumerate(form.photos.data):
                 file.save(os.path.join(f'static/albums/{album.translit_name}',  f'{i}.jpg'))
             session.add(album)
             session.commit()
+            flash('Успех')
             return redirect(f'/albums/{album.translit_name}')
         return render_template('album_form.html', title='Загрузить альбом', form=form, albums=get_albums())
     abort(404)
@@ -180,7 +189,7 @@ def upload():
 
 @app.route('/albums/edit/<name>', methods=['GET', 'POST'])
 def edit_album(name):
-    if current_user.id == 1:
+    if current_user.is_authenticated and current_user.id == 1:
         form = AlbumForm()
         session = db_session.create_session()
         album = session.query(Album).filter(Album.translit_name == name).first()
@@ -192,6 +201,7 @@ def edit_album(name):
                 file.save(os.path.join(f'static/albums/{album.translit_name}', f'{i}.jpg'))
             album.lenght += len(form.photos.data)
             session.commit()
+            flash('Успех')
             return redirect(f'/albums/{name}')
         form.name.data = album.name
         return render_template('album_form.html', title='Редактировать альбом', form=form, albums=get_albums())
@@ -200,15 +210,75 @@ def edit_album(name):
 
 @app.route('/albums/delete/<name>')
 def delete_album(name):
-    if current_user.id == 1:
+    if current_user.is_authenticated and current_user.id == 1:
         session = db_session.create_session()
         album = session.query(Album).filter(Album.translit_name == name).first()
         if album:
             shutil.rmtree(os.path.join('static/albums', name))
             session.delete(album)
             session.commit()
+            flash('Успех')
             return redirect('/')
     abort(404)
+
+
+@app.route('/pricelist', methods=['GET', 'POST'])
+def booking():
+    session = db_session.create_session()
+    form = BookingForm()
+    if form.validate_on_submit():
+        book = Booking()
+        book.name = form.name.data
+        book.surname = form.surname.data
+        book.tel = form.tel.data
+        book.type = form.type.data
+        book.date = form.date.data
+        book.content = form.content.data
+        session.add(book)
+        session.commit()
+        flash('Успешно отправлено')
+        return redirect('/pricelist')
+    if current_user.is_authenticated and current_user.id == 1:
+        books = session.query(Booking).all()[::-1]
+        return render_template('booking.html', title='Услуги и цены', books=books, albums=get_albums())
+    return render_template('booking.html', title='Услуги и цены', form=form, albums=get_albums())
+
+
+@app.route('/booking/delete/<int:id>')
+def book_delete(id):
+    if current_user.is_authenticated and current_user.id == 1:
+        session = db_session.create_session()
+        book = session.query(Booking).filter(Booking.id == id).first()
+        session.delete(book)
+        session.commit()
+        flash('Успех')
+        return redirect('/pricelist')
+    abort(404)
+
+
+@app.route('/about')
+def about():
+    return render_template('about.html', title='Обо мне', albums=get_albums())
+
+
+@app.route('/contacts')
+def contacts():
+    return render_template('contacts.html', title='Контакты', albums=get_albums())
+
+
+def feel_db():
+    session = db_session.create_session()
+    for i in range(1, 6):
+        review = Review()
+        review.user_id = 1
+        review.user_name = session.query(User).filter(User.id == 1).first().name
+        review.user_surname = session.query(User).filter(User.id == 1).first().surname
+        review.rating = i
+        review.content = 'Lorem ipsum dolor sit amet consectetur, adipisicing elit. Quisquam ipsam fugiat natus ' \
+                         'tempora molestias illo quis provident quod, maiores doloremque?'
+        review.date = datetime.date.today()
+        session.add(review)
+    session.commit()
 
 
 if __name__ == '__main__':
